Lang

+++ src/Infrastructure/Infrastructure/Migrations/20260324060613_Languages.cs
@@ -14,7 +14,7 @@ namespace GamaEdtech.Infrastructure.Migrations
                 name: "Languages",
                 columns: table => new
                 {
-                    Id = table.Column<long>(type: "bigint", nullable: false)
+                    Id = table.Column<long>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(100)", maxLength: 100, nullable: false),
                     Code = table.Column<string>(type: "varchar(20)", maxLength: 20, nullable: false),
